requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic
python-dotenv
requests
huggingface_hub
